# ui/openapi.json
{
  "components": {
    "schemas": {
...
      "Connection": {
        "properties": {
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "username": {
            "default": "",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "Connection",
        "type": "object"
      },
...
      "Group": {
        "description": "A Music Assistant sync group, or a solo speaker as a group of one (``id`` is null).",
        "properties": {
          "avg_volume": {
            "default": 0,
            "title": "Avg Volume",
            "type": "integer"
          },
          "external_count": {
            "default": 0,
            "title": "External Count",
            "type": "integer"
          },
          "external_members": {
            "description": "Members on other bridges or other player providers.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "External Members",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/GroupMember"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "playing": {
            "default": false,
            "title": "Playing",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "members",
          "avg_volume",
          "playing",
          "external_members",
          "external_count"
        ],
        "title": "Group",
        "type": "object"
      },
...
      "HaSignIn": {
        "properties": {
          "code": {
            "default": "",
            "title": "Code",
            "type": "string"
          },
          "ma_url": {
            "title": "Ma Url",
            "type": "string"
          },
          "password": {
            "default": "",
            "title": "Password",
            "type": "string"
          },
          "step": {
            "default": "init",
            "enum": [
              "init",
              "mfa"
            ],
            "title": "Step",
            "type": "string"
          },
          "username": {
            "default": "",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "ma_url"
        ],
        "title": "HaSignIn",
        "type": "object"
      },
...
      "HaSilentIn": {
        "properties": {
          "ha_token": {
            "title": "Ha Token",
            "type": "string"
          },
          "ma_url": {
            "title": "Ma Url",
            "type": "string"
          }
        },
        "required": [
          "ha_token",
          "ma_url"
        ],
        "title": "HaSilentIn",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "description": "What the job does, e.g. bluetooth.scan, bluetooth.pairing, updates.check.",
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "additionalProperties": true,
            "title": "Progress",
            "type": "object"
          },
          "result": {
            "title": "Result"
          },
          "status": {
            "default": "running",
            "enum": [
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the job acts on (a MAC, a device id…).",
            "title": "Subject"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "created_at",
          "updated_at",
          "subject",
          "progress",
          "result",
          "error"
        ],
        "title": "Job",
        "type": "object"
      },
...
      "PasswordSignIn": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "url": {
            "default": "",
            "description": "Empty: find the server (saved URL, Sendspin peer, mDNS).",
            "title": "Url",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "PasswordSignIn",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "An error response. Use cases may add fields (e.g. ``remediation``).",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation specific to this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Field-level validation errors.",
            "title": "Errors"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "URI reference identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "QueueCommandIn": {
        "properties": {
          "action": {
            "enum": [
              "next",
              "previous",
              "play",
              "pause",
              "shuffle",
              "repeat",
              "seek"
            ],
            "title": "Action",
            "type": "string"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "syncgroup_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Syncgroup Id"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "action"
        ],
        "title": "QueueCommandIn",
        "type": "object"
      },
      "QueueCommandOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "ma_now_playing": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Predicted state, shown at once.",
            "title": "Ma Now Playing"
          },
          "op_id": {
            "title": "Op Id",
            "type": "string"
          },
          "queue_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Id"
          },
          "syncgroup_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Syncgroup Id"
          }
        },
        "required": [
          "job",
          "op_id"
        ],
        "title": "QueueCommandOut",
        "type": "object"
      },
...
      "ReloadOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "monitor_reloaded": {
            "title": "Monitor Reloaded",
            "type": "boolean"
          }
        },
        "required": [
          "monitor_reloaded",
          "job"
        ],
        "title": "ReloadOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Bridges Music Assistant's Sendspin players to Bluetooth speakers. One API for the web UI, Home Assistant and automation.",
    "title": "Sendspin Bluetooth Bridge",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "operationId": "list_groups_api_v1_groups_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  },
                  "title": "Response List Groups Api V1 Groups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Groups and their members",
        "tags": [
          "bridge"
        ]
      }
    },
...
    "/api/v1/music-assistant": {
      "get": {
        "operationId": "connection_api_v1_music_assistant_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Connection Api V1 Music Assistant Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Connection state",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/artwork": {
      "get": {
        "operationId": "artwork_api_v1_music_assistant_artwork_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "title": "Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sig",
            "required": false,
            "schema": {
              "default": "",
              "title": "Sig",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/*": {}
            },
            "description": "The image"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Signed artwork, same-origin",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/debug": {
      "get": {
        "operationId": "debug_api_v1_music_assistant_debug_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Debug Api V1 Music Assistant Debug Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Caches, groups and live queues",
        "tags": [
          "music-assistant",
          "diagnostics"
        ]
      }
    },
    "/api/v1/music-assistant/discovery": {
      "post": {
        "operationId": "discovery_api_v1_music_assistant_discovery_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Find Music Assistant servers",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/groups": {
      "get": {
        "operationId": "groups_api_v1_music_assistant_groups_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Groups Api V1 Music Assistant Groups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Music Assistant sync groups",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/groups/refresh": {
      "post": {
        "operationId": "refresh_groups_api_v1_music_assistant_groups_refresh_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Re-read sync groups",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/now-playing": {
      "get": {
        "description": "For one Music Assistant sync group with ``syncgroup_id``; without it, the first known group.",
        "operationId": "now_playing_api_v1_music_assistant_now_playing_get",
        "parameters": [
          {
            "in": "query",
            "name": "syncgroup_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Syncgroup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Now Playing Api V1 Music Assistant Now Playing Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "What is playing",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/queue-commands": {
      "post": {
        "operationId": "queue_command_api_v1_music_assistant_queue_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueCommandIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueCommandOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Next, previous, shuffle, repeat, seek",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/reload": {
      "post": {
        "operationId": "reload_api_v1_music_assistant_reload_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Reconnect with the saved credentials",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/session": {
      "post": {
        "operationId": "sign_in_api_v1_music_assistant_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordSignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Sign in with Music Assistant credentials",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/session/ha": {
      "post": {
        "operationId": "sign_in_ha_api_v1_music_assistant_session_ha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HaSignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HaSignInResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Sign in through Home Assistant (MA as an HA add-on)",
        "tags": [
          "music-assistant"
        ]
      }
    },
    "/api/v1/music-assistant/session/ha-silent": {
      "post": {
        "operationId": "sign_in_ha_silent_api_v1_music_assistant_session_ha_silent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HaSilentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Server error"
          }
        },
        "summary": "Add-on mode: sign in with the user's HA token",
        "tags": [
          "music-assistant"
        ]
      }
    },
...
  }
}